python/02_fastapi/14_todoapp_auth/requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib
bcrypt==3.2.0
python-multipart